continue example
An example of using continue: An airline ticketing system must calculate the total cost of all purchased tickets. Tickets for children under 3 years old are free. We can use a while loop to iterate through the list of passengers and calculate the total cost of their tickets. In this case, the continue statement can be used to skip the children.